refactor: implement GetLatestStatusReport db func and update timeout detector
