+++ backend/pkg/commons/db2/db.go
@@ -43,6 +43,7 @@ type LastBlocksStoreWriter interface {
 
 type Monitoring interface {
 	SaveNewStatusReport(status StatusReport) error
+	GetLatestStatusReport() ([]Victims, error)
 	GetEmitters() ([]string, error)
 	GetVDLatestEpochTs() (time.Time, error)
 	GetVDRollingEpochEnd(rolling string) (uint64, error)
